Convert build system into using cmake.
git-svn-id: svn://svn.code.sf.net/p/jack-keyboard/code/trunk@21 1fa2bf75-7d80-4145-9e94-f9b4e25a1cb2

+++ src/pianokeyboard.c
@@ -217,11 +217,11 @@ key_binding(PianoKeyboard *pk, const char *key)
 	if (!found)
 		return (-1);
 
-	return ((int)note);
+	return ((long)note);
 }
 
 static void
-bind_key(PianoKeyboard *pk, const char *key, int note)
+bind_key(PianoKeyboard *pk, const char *key, long note)
 {
 	assert(pk->key_bindings != NULL);
